+++ src/org/supremica/automata/IO/AutomataToXml.java
@@ -219,7 +219,7 @@ public class AutomataToXml
                 
                 if (includeCost)
                 {
-                    int value = state.getCost();
+                    double value = state.getCost();
                     
                     if (value != State.UNDEF_COST)
                     {
